requirements.txt
alabaster==0.7.13
amqp==5.1.1
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
astroid==2.15.6
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
billiard==4.1.0
binaryornot==0.4.4
black==23.7.0
cachetools==5.3.1
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.3.0
coverage==7.2.7
crispy-bootstrap5==0.7
cron-descriptor==1.4.0
cryptography==41.0.3
cssbeautifier==1.14.9
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
Django==4.2.4
django-allauth==0.54.0
django-anymail==10.1
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-coverage-plugin==3.1.0
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-extensions==3.2.3
django-model-utils==4.3.1
django-redis==5.3.0
django-storages==1.13.2
django-stubs==4.2.3
django-stubs-ext==4.2.2
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.6.0
djangorestframework-stubs==3.14.2
djlint==1.32.1
docutils==0.19
drf-spectacular==0.26.4
EditorConfig==0.12.3
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
factory-boy==3.3.0
Faker==19.3.0
filelock==3.12.2
flake8==6.1.0
flake8-isort==6.0.0
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.3.3
google-cloud-secret-manager==2.16.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.57.0
grpcio-status==1.57.0
gunicorn==21.2.0
hiredis==2.2.3
html-tag-names==0.1.2
html-void-elements==0.1.0
httplib2==0.22.0
identify==2.5.26
idna==3.4
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.14.0
isort==5.12.0
jedi==0.19.0
Jinja2==3.1.2
jsbeautifier==1.14.9
json5==0.9.14
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.1
lazy-object-proxy==1.9.0
livereload==2.6.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mypy==1.4.1
mypy-extensions==1.0.0
mysqlclient==2.2.0
nodeenv==1.8.0
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.2.0
pre-commit==3.3.3
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.0
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg-c==3.1.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
pylint==2.17.5
pylint-celery==0.3
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
PyMySQL==1.1.0
pyparsing==3.1.1
pytest==7.4.0
pytest-django==4.5.2
pytest-sugar==0.9.7
python-barcode==0.15.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
sentry-sdk==1.29.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==6.2.1
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.5
sphinxcontrib-devhelp==1.0.3
sphinxcontrib-htmlhelp==2.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.4
sphinxcontrib-serializinghtml==1.1.6
sqlparse==0.4.4
stack-data==0.6.2
termcolor==2.3.0
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.2
tqdm==4.66.0
traitlets==5.9.0
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.11
types-requests==2.31.0.2
types-urllib3==1.26.25.14
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
virtualenv==20.24.2
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
Werkzeug==2.3.6
whitenoise==6.5.0
wrapt==1.15.0
